refactor: use same restoreJsonRpcError function for renderer worker and extension host worker
