+++ packages/extension-host-worker/src/parts/ErrorType/ErrorType.js
@@ -1,0 +1,4 @@
+export const DomException = 'DOMException'
+export const ReferenceError = 'ReferenceError'
+export const SyntaxError = 'SyntaxError'
+export const TypeError = 'TypeError'
