+++ packages/extension-host-worker/src/parts/RestoreJsonRpcError/RestoreJsonRpcError.js
