--- packages/renderer-worker/src/parts/Errors/Errors.js
@@ -1,6 +1,0 @@
-export class JsonRpcError extends Error {
-  constructor(message) {
-    super(message)
-    this.name = 'JsonRpcError'
-  }
-}
