+++ packages/renderer-worker/src/parts/JsonRpcError/JsonRpcError.js
@@ -1,0 +1,6 @@
+export class JsonRpcError extends Error {
+  constructor(message) {
+    super(message)
+    this.name = 'JsonRpcError'
+  }
+}
